refactor: rename AppHead
